Update contact.h
Explanation:
Additional Member Variable:

linkedCoil: Pointer to the linked coil element. This allows the contact to be linked to a coil element.
Method for Linking to Coil:

linkToCoil: Links the contact to a coil element by storing a shared pointer to the coil.
Integration with CoilElement Class:

We've added a forward declaration of the CoilElement class. This is necessary for using CoilElement in the Contact class.
With this update, a contact element can be linked to a coil element by calling the linkToCoil method. This association allows the contact to interact with the coil element, typically in the context of controlling it based on the contact's state.